separate go plugin into multiple files
